Avoid showing re-sent attachments twice in music browser
Outgoing attachments sent from my phone (from iTunes) have an assetURL,
these should be shown in the music browser. Outgoing attachments that do not
have an assetURL are re-sent from an incoming attachment that is already shown
in the music browser, so we do not need to show it twice.