Fix umlaut confusion by not using umlauts in filenames
Due to a bug in Git <1.8.2 on Mac OS X, the files with umlauts in the names
got added twice -- once with the composed umlaut character (0xc3bf) and once
with the decomposed form (0x6fcc88). The last commit removes one version, this
commit renames all files with umlauts in the name.

Welcome to 2015. :)